Added assimp and sponza test model.
I am having this weird error where sometimes I'll get "undefined symbols" and "cannot find threeDplay.dylib" when trying to build. Cmake seems to work fine, this is a problem in make/cc. I did have to update my version of CMake to build assimp. 3.20 => 3.30  . I have a feeling this is either CLion messing things up accidentally, or something to do with the order of compliation. Problem is always fixed when i remove the build directory and do a full, clean build. But that takes super long. so..... But in general, the program builds and runs and assimp is able to open the sponza scene)